requirements.txt
aiomysql==0.0.21
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
arrow==1.3.0
bcrypt==4.1.2
binaryornot==0.4.4
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.5.0
cryptography==42.0.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Faker==20.1.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.2
jinja2-time==0.2.0
jose==1.0.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.0.22
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
paramiko==3.4.0
passlib==1.7.4
pillow==10.2.0
poyo==0.5.0
protobuf==3.14.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
PyMySQL==0.9.3
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
razorpay==1.4.1
requests==2.31.0
rich==13.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlmodel==0.0.14
sshtunnel==0.4.0
starlette==0.27.0
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.24.0
